pyproject.toml
[project]
name = "docmind-ai"
version = "1.0.0"
description = "Custom AI Chatbot Builder for Businesses"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "chromadb>=0.5.20",
    "langchain>=0.3.10",
    "langchain-community>=0.3.10",
    "langchain-chroma>=0.1.4",
    "sentence-transformers>=3.3.1",
    "pypdf2>=3.0.1",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "ollama>=0.4.4",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "redis>=5.2.0",
    "slowapi>=0.1.9",
    "loguru>=0.7.2",
    "httpx>=0.28.1",
    "email-validator>=2.3.0",
    "pypdf>=6.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "ruff>=0.8.4",
]

[tool.black]
line-length = 100
target-version = ['py313']

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
